Fix aliasing when generating models from a local checkpoint folder with compression
